Templatize storyboard mocking functions
In doing this, we eliminate the need to explicitly implement
error handling to keep users from mocking storyboard elements
with anything other than UIViewController classes. The compiler
does this for us now.

This restricts the users to mocking only UIViewControllers on
storyboards for now. It will have to be reconsidered should
there be a need to allow mocking of other types of objects on
storyboards.

[#134637307]