Advance run loop in window set up helper methods
When setting up navigation controllers as the new root view controllers
for windows in test, the UI threads did not seem to complete their setup
process as they do for vanilla `UIViewControllers`. This caused any
views nested within the `UINavigationController` to be unable to take
first responder. This, in turn, causes anyone using Fleet's helpers for
text input views to fail instantly if they chose to set their view
controllers in-test to live with navigation controllers -- a recommended
practice in any case that a view controller lives in a nav controller in
production code.

Also renamed the root window view helper method and added a convenience
version to set a view controller up in a navigation controller for you
automatically.